+++ lib/ember-cli/html_page.rb
@@ -1,0 +1,12 @@
+module EmberCLI
+  class HtmlPage
+    def initialize(options)
+      @content = options.fetch(:content)
+      @asset_resolver = options.fetch(:asset_resolver)
+    end
+
+    def render
+      @asset_resolver.resolve_urls(@content)
+    end
+  end
+end