+++ spec/features/user_views_ember_app_spec.rb
@@ -1,6 +1,12 @@
 feature "User views ember app", :js do
-  scenario "from root" do
+  scenario "with asset helpers" do
     visit root_path
+
+    expect(page).to have_text "Welcome to Ember"
+  end
+
+  scenario "with index helper" do
+    visit root_path(index_html: true)
 
     expect(page).to have_text "Welcome to Ember"
   end